Add `Pf::USet32#usize`, `Pf::USet32#prefix`
- `Pf::USet32#usize` returns the original cardinality before it's cast
to_i. This is useful because one may want to know the size of a set of
more than 2**31 elements.
- `Pf::USet32#prefix`, bitwise, returns the number of trailing ones
really quickly (in a matter of a dozen nanoseconds for sets of >10000
consecutive elements [such sets are a worst-case]). Set-wise, it returns
the length of the consecutive run of integers starting from 0 in the
set. This will be very useful for Wirewright's `Term::Dict`, which would
really like to have something like this for its itemspart. Right now we
do an expensive merging process, moving items between the itemspart and
the pairspart. With this and a few other upcoming methods, we'll be able
to map arbitrary u32 keys in dict to an offset and store them in a
contiguous, persistent array. This method, then, can be used for
`Term::Dict#itemsize`.